Revert "Merge pull request #276 from PetaVision/pv_stream-removed"
This reverts commit 421dbbe5c4e74e9acb111eae4cd0276697bbf363, reversing
changes made to a1e31d600c86f6f4d8ef0f963387da92756f2b58.

The merge should have been into develop, not master.

+++ src/structures/Buffer.tpp
@@ -1,5 +1,5 @@
 #include "utils/PVLog.hpp"
-//#include "utils/conversions.hpp"
+//#include "utils/conversions.h"
 
 #include <cmath>
 #include <cstring>
@@ -52,7 +52,7 @@ void Buffer<T>::set(int k, T value) {
 template <class T>
 void Buffer<T>::set(const std::vector<T> &vector, int width, int height, int features) {
    FatalIf(
-         (int)vector.size() != width * height * features,
+         vector.size() != width * height * features,
          "Invalid vector size: Expected %d elements, vector contained %d elements.\n",
          width * height * features,
          vector.size());
@@ -92,8 +92,8 @@ void Buffer<T>::grow(int newWidth, int newHeight, enum Anchor anchor) {
    if (newWidth <= getWidth() && newHeight <= getHeight()) {
       return;
    }
-   newWidth    = std::max(newWidth, getWidth());
-   newHeight   = std::max(newHeight, getHeight());
+   newWidth  = std::max(newWidth, getWidth());
+   newHeight = std::max(newHeight, getHeight());
    int offsetX = getAnchorX(anchor, getWidth(), newWidth);
    int offsetY = getAnchorY(anchor, getHeight(), newHeight);
    Buffer bigger(newWidth, newHeight, getFeatures());
